refactor(create-pyric): split the chat page into components
ChatPage rendered the sidebar list, the presence bar, the conversation
header, and the message thread as inline JSX, so the only component in
the tree was ChatPage itself. The runtime chip attributes a listener to
the component that opened it and outlines the subtree that re-rendered
after a delivery, and with one component there was nothing smaller to
outline.

The four regions are now components beside the page: ConversationList
with a ConversationItem per entry, ConversationHeader, MessageThread
with a MessageBubble per message, and PresenceBar. Each takes the data
and callbacks it needs as props, and each takes its region ref as a
plain ref prop, which React 19 passes to a function component without
forwardRef. ChatPage keeps the state, the effects, useListenerOwner, and
the ownerFor call that pairs the captured owner with the region element,
so every listener is opened and attributed exactly as before. The
markup, class names, aria attributes, and handlers are unchanged.

Message helpers that only the thread uses moved with it, and the two
shared formatters live in chat-format.ts.
